Rework PrimeCloudSync keyless flow
