+++ packages/kit-bg/src/apis/BackgroundApi.ts
@@ -382,6 +382,16 @@ class BackgroundApi extends BackgroundApiBase implements IBackgroundApi {
     return value;
   }
 
+  get serviceKeylessCloudSync() {
+    const Service =
+      require('../services/ServiceKeylessCloudSync') as typeof import('../services/ServiceKeylessCloudSync');
+    const value = new Service.default({
+      backgroundApi: this,
+    });
+    Object.defineProperty(this, 'serviceKeylessCloudSync', { value });
+    return value;
+  }
+
   get serviceQrWallet() {
     const Service =
       require('../services/ServiceQrWallet') as typeof import('../services/ServiceQrWallet');
