+++ packages/kit-bg/src/services/ServiceKeylessCloudSync/index.ts
@@ -1,0 +1,3 @@
+import ServiceKeylessCloudSync from './ServiceKeylessCloudSync';
+
+export default ServiceKeylessCloudSync;
